feat: implementando a mudança de status

+++ src/app/adm/editais/[id]/(actions)/delete-event-action.ts
@@ -1,10 +1,13 @@
-'use server'
+"use server";
 
-import { revalidatePath } from 'next/cache'
-import { kyClient } from '@/infra/external/http/ky-client/api'
+import { revalidatePath } from "next/cache";
+import { kyClient } from "@/infra/external/http/ky-client/api";
 
-export async function deleteEventAction(stepId: number) {
-  await kyClient.delete(`steps/${stepId}`)
+export async function deleteEventAction(formData: FormData) {
+  const stepId = Number(formData.get("stepId"));
+  if (!stepId) return;
 
-  revalidatePath('/adm/editais/[id]', 'page')
+  await kyClient.delete(`steps/${stepId}`);
+
+  revalidatePath("/adm/editais/[id]", "page");
 }
